Use anyhow::Error in place of custom type
anyhow::Error has several features on top of RoperatorError including
use of `try!` and `?`, chaining, and adding context to errors. it's also
a defacto standard error type at the moment in the broader community.

on the user side, the only change required is to use anyhow::Error and
switch `Box::new(err)` to `Error::new(err)`.

all tests are passing and open to thoughts on how to make the change
less impactful.